+++ NEWS.md
@@ -2,4 +2,6 @@
 
 ## New features
 
+* Renamed `findTCGAworkspaces` to `selectTCGAworkspace`; the latter allows for
+an interactive selection of TCGA data workspaces.
 * Package submitted to Bioconductor!